just one problem time limit

+++ test.py
@@ -1,7 +1,7 @@
 n = int(input())
 while n>0:
-    num = int(input())
-    while(num>0 and num!=1):
+    num = float(input())
+    while (num>0 and num!=1):
         if(num%2==0):
             num=num/2
         else:
